requirements.txt
streamlit
pymongo
bcrypt
python-dotenv
sentence-transformers
faiss-cpu
arxiv
requests
google-generativeai
